requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.1.2
google-generativeai==0.3.1
psycopg2-binary==2.9.9
gunicorn==21.2.0